refactor(gallery): remove debugging delay and streamline agent data fetching
- Eliminated the 5-second debugging delay from GalleryPage and individual agent pages to enhance performance.
- Refactored the landing page to utilize AgentsGallery for displaying agents, improving code organization and readability.
